channels: better header mgmt
