+++ ui/package.json
@@ -94,7 +94,7 @@
     "@types/prosemirror-view": "^1.23.3",
     "@urbit/api": "^2.2.0",
     "@urbit/aura": "^0.4.0",
-    "@urbit/http-api": "^2.4.3-debug",
+    "@urbit/http-api": "^2.4.4-debug",
     "any-ascii": "^0.3.1",
     "big-integer": "^1.6.51",
     "browser-cookies": "^1.2.0",
