+++ ui/src/diary/DiaryChannel.tsx
@@ -2,7 +2,6 @@ import React, { useCallback, useEffect, useState } from 'react';
 import { Outlet, useLocation, useNavigate, useParams } from 'react-router';
 import bigInt from 'big-integer';
 import { Virtuoso } from 'react-virtuoso';
-import { Link } from 'react-router-dom';
 import * as Toast from '@radix-ui/react-toast';
 import Layout from '@/components/Layout/Layout';
 import {
@@ -17,16 +16,9 @@ import {
   useDiaryDisplayMode,
   useDiaryPerms,
 } from '@/state/diary';
-import {
-  DiarySetting,
-  setChannelSetting,
-  useDiarySettings,
-  useDiarySortMode,
-  usePutEntryMutation,
-} from '@/state/settings';
-import ChannelHeader from '@/channels/ChannelHeader';
+import { useDiarySortMode } from '@/state/settings';
 import useDismissChannelNotifications from '@/logic/useDismissChannelNotifications';
-import { DiaryDisplayMode, DiaryLetter } from '@/types/diary';
+import { DiaryLetter } from '@/types/diary';
 import DiaryGridView from '@/diary/DiaryList/DiaryGridView';
 import useRecentChannel from '@/logic/useRecentChannel';
 import {
@@ -35,11 +27,11 @@ import {
   isChannelJoined,
 } from '@/logic/utils';
 import useAllBriefs from '@/logic/useAllBriefs';
-import AddIcon16 from '@/components/icons/Add16Icon';
 import { useLastReconnect } from '@/state/local';
 import DiaryListItem from './DiaryList/DiaryListItem';
 import useDiaryActions from './useDiaryActions';
 import DiaryChannelListPlaceholder from './DiaryChannelListPlaceholder';
+import DiaryHeader from './DiaryHeader';
 
 function DiaryChannel() {
   const [joining, setJoining] = useState(false);
@@ -60,10 +52,6 @@ function DiaryChannel() {
     ? isChannelJoined(nest, briefs)
     : true;
   const lastReconnect = useLastReconnect();
-  const { mutate } = usePutEntryMutation({
-    bucket: 'diary',
-    key: 'settings',
-  });
   const needsLoader = letters.size === 0;
 
   const joinChannel = useCallback(async () => {
@@ -91,31 +79,12 @@ function DiaryChannel() {
     time: newNote || '',
   });
 
-  const settings = useDiarySettings();
   // for now sortMode is not actually doing anything.
   // need input from design/product on what we want it to actually do, it's not spelled out in figma.
   const displayMode = useDiaryDisplayMode(chFlag);
   const sortMode = useDiarySortMode(chFlag);
 
-  const setDisplayMode = async (view: DiaryDisplayMode) => {
-    await useDiaryState.getState().viewDiary(chFlag, view);
-  };
-
-  const setSortMode = (
-    setting: 'time-dsc' | 'quip-dsc' | 'time-asc' | 'quip-asc'
-  ) => {
-    const newSettings = setChannelSetting<DiarySetting>(
-      settings,
-      { sortMode: setting },
-      chFlag
-    );
-    mutate({
-      val: JSON.stringify(newSettings),
-    });
-  };
-
   const perms = useDiaryPerms(chFlag);
-
   const canWrite = canWriteChannel(perms, vessel, group?.bloc);
   const canRead = channel
     ? canReadChannel(channel, vessel, group?.bloc)
@@ -201,28 +170,13 @@ function DiaryChannel() {
       className="flex-1 bg-gray-50"
       aside={<Outlet />}
       header={
-        <ChannelHeader
-          isDiary
+        <DiaryHeader
           flag={flag}
           nest={nest}
-          showControls
-          displayMode={displayMode}
-          setDisplayMode={setDisplayMode}
-          sortMode={sortMode}
-          setSortMode={setSortMode}
-        >
-          {canWrite ? (
-            <Link
-              to="edit"
-              className={
-                'small-button shrink-0 bg-blue px-1 text-white sm:px-2'
-              }
-            >
-              <AddIcon16 className="h-4 w-4 sm:hidden" />
-              <span className="hidden sm:inline">Add Note</span>
-            </Link>
-          ) : null}
-        </ChannelHeader>
+          canWrite={canWrite}
+          display={displayMode}
+          sort={sortMode}
+        />
       }
     >
       <Toast.Provider>
